mv some basic functions to fem_ele_vec_utils.cpp
